Fix artifacts having unintended effects and segfaults.
Another bugfix backport.  Credit goes to AmyBSOD.

+++ src/do_name.c
@@ -296,7 +296,7 @@ register struct obj *obj;
 {
 	char buf[BUFSZ], qbuf[QBUFSZ];
 	const char *aname;
-	short objtyp;
+	int objtyp;
 
 	Sprintf(qbuf, "What do you want to name %s %s?",
 		is_plural(obj) ? "these" : "this", xname(obj));
